pyproject.toml
[project]
name = "instachef"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "instaloader>=4.15.1",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.0.0",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.9",
]
